Added processIntents and embedded image/video examples
Added a processIntents function that intercepts messages between the
Watson Service and the interface, processIntents will find the highest
confidenence intent returned from Watson and check if any custom
messages should be returned from the node.js server. In this case we've
implemented hardcoded examples of imbedded images and videos for
demonstration.